Add ErrorResponseDto class for error response handling

+++ src/main/java/nycto/homeservices/dto/commentDto/CommentResponseDto.java
@@ -5,6 +5,6 @@ public record CommentResponseDto(
         Long customerId,
         Long orderId,
         String content,
-        Integer rating
+        Double rating
 ) {
 }
